refactor(fspy-shm): split platform implementations
